backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==3.2.2
certifi==2026.1.4
cffi==2.0.0
click==8.1.8
cryptography==46.0.5
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.109.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.1.0
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.21.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.2
pycparser==2.23
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.4.0
tqdm==4.67.3
typing_extensions==4.15.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
